mathml/relations/html5-tree/css-inline-style-interface.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathML 'ElementCSSInlineStyle` Mixin Tests</title>
    <link
      rel="help"
      href="https://mathml-refresh.github.io/mathml-core/#dom-and-javascript"
    />
    <style>
        mi { background-color: red; }
      }
    </style>
    <meta name="assert" content="MathMLElements incorporate a functional ElementCSSInlineStyle interface" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

  </head>
  <body>
    <span>This tests the presence and functionality of features of the `ElementCSSInlineStyle` interface for MathMLElements</span>
    <math>
      <mi>E</mi>
    </math>
  </body>
  <script>

    let el = document.querySelector("mi");
    let mathEl = document.querySelector("math");

    test(function() {
      el.style.backgroundColor = "blue";
      assert_equals(
        getComputedStyle(el).backgroundColor,
        "rgb(0, 0, 255)",
        "The applied background should be blue."
      );
    }, "The style property should be present and be functional.");

  </script>
</html>
